pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snapshot-query"
version = "0.1.1"
description = "Efficiently query and analyze snapshot log files generated by cursor-ide-browser"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Cursor IDE Team", email = "support@cursor.com"},
]
keywords = ["cursor", "browser", "snapshot", "query", "accessibility", "mcp", "pydantic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
snapshot-query = "snapshot_query.cli:main"
snapshot-query-mcp = "snapshot_query.mcp_server:run_server"

[project.urls]
Homepage = "https://github.com/cursor-ide/snapshot-query"
Documentation = "https://github.com/cursor-ide/snapshot-query#readme"
Repository = "https://github.com/cursor-ide/snapshot-query"

[tool.hatch.build.targets.wheel]
packages = ["snapshot_query"]
